requirements/requirements-general.txt
# Зависимости для общего окружения
# Установка: pip install -r requirements-general.txt

jupyter
ollama
numpy==2.3
pandas
scikit-learn
matplotlib
seaborn
scipy
openpyxl
pyarrow
shap
statsmodels
catboost
xgboost
imbalanced-learn
h5py
geopandas
gpxpy
shapely
mapclassify
folium
rasterio
contextily
osmnx
librosa
opencv-python
ultralytics
python-docx
pymupdf
tqdm
requests
fastapi
aiogram
streamlit
streamlit-folium
langchain
langchain-community
faiss-cpu
